requirements.txt
psycopg2-binary
tabulate
python-dotenv